requirements.txt
contourpy==1.3.2
cycler==0.12.1
fonttools==4.58.0
kiwisolver==1.4.8
matplotlib==3.10.3
numpy==2.2.6
packaging==25.0
pandas==2.2.3
pillow==11.2.1
py4j==0.10.9.7
pyparsing==3.2.3
pyspark==3.5.5
python-dateutil==2.9.0.post0
pytz==2025.2
setuptools==80.8.0
six==1.17.0
tzdata==2025.2
